requirements.txt
streamlit>=1.33
pandas>=2.0
sqlalchemy>=2.0
pydantic>=2.5
python-dotenv>=1.0
pdfplumber>=0.11
python-docx>=1.1
matplotlib>=3.7
